% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gris.R
\name{as.gris.triangulation}
\alias{as.gris.triangulation}
\title{as.gris}
\usage{
\method{as.gris}{triangulation}(x, ...)
}
\arguments{
\item{x}{triangulation (package RTriangle)}

\item{type}{mesh, triangle polygons, or triangle edges}
}
\description{
as.gris
}

