% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gris.R
\name{bld}
\alias{bld}
\title{Generate a vertex/branch/object table from Spatial polygons}
\usage{
bld(x, ...)
}
\arguments{
\item{x}{SpatialPolygonsDataFrame}

\item{...}{ignored}
}
\value{
list of \code{v} vertex, \code{b} branch and \code{o} object tables
}
\description{
Generate a vertex/branch/object table from Spatial polygons
}

