% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmu.R
\docType{data}
\name{cmu}
\alias{cmu}
\title{CMU banner}
\format{A gris object composed of five line objects.}
\source{
This was generated by tracing the image used on the Triangle site
in Manifold GIS, then exporting that to MapInfo file (.mif) and reading it in with rgdal. \url{https://www.cs.cmu.edu/~quake/triangle.demo.html}
library(rgdal); library(gris);  cmu <- gris(readOGR("CMU.mif", "CMU"))
}
\usage{
cmu
}
\description{
A gris dataset of the example CMU banner used by Shewchuk to illustrate the features of Triangle.
}
\keyword{datasets}

