% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gris.R
\name{dv}
\alias{dv}
\title{Generate a vertex/branch/object table from Spatial polygons}
\usage{
dv(x, ...)
}
\arguments{
\item{x}{SpatialPolygonsDataFrame}

\item{...}{ignored}
}
\value{
vertex/branch/object table suitable for \code{pl}
}
\description{
Generate a vertex/branch/object table from Spatial polygons
}

