% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign.R
\name{as.SpatialPolygonsDataFrame}
\alias{as.SpatialPolygonsDataFrame}
\alias{as.SpatialPolygonsDataFrame.gris}
\title{Convert gris to Spatial}
\usage{
as.SpatialPolygonsDataFrame(x, ...)

\method{as.SpatialPolygonsDataFrame}{gris}(x, ...)
}
\arguments{
\item{x}{gris object}

\item{...}{ignored}
}
\description{
Spatial classes are defined in the \link{sp} package.
}

