% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocentric.r
\name{geocentric}
\alias{geocentric}
\title{Geocentric coordinates.}
\usage{
geocentric(x, crs = "+proj=geocent +ellps=WGS84")
}
\arguments{
\item{x}{table of long, lat, and optionally "height"}

\item{crs}{}

\item{exag}{exaggeration to apply to height values (added to radius)}
}
\value{
tibble
}
\description{
Transform to geocentric coordinates.
}

