% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proj4.R
\name{xmin}
\alias{local_triangulate}
\alias{xmax}
\alias{xmax.default}
\alias{xmax.gris}
\alias{xmin}
\alias{xmin.default}
\alias{xmin.gris}
\alias{xrange}
\alias{xrange.default}
\alias{ymax}
\alias{ymax.default}
\alias{ymax.gris}
\alias{ymin}
\alias{ymin.default}
\alias{ymin.gris}
\alias{yrange}
\alias{yrange.default}
\title{min/max for gris}
\usage{
xmin(x)

xmax(x)

ymin(x)

ymax(x)

\method{xmin}{default}(x)

\method{xmax}{default}(x)

\method{ymin}{default}(x)

\method{ymax}{default}(x)

\method{xmin}{gris}(x)

\method{xmax}{gris}(x)

\method{ymin}{gris}(x)

\method{ymax}{gris}(x)

xrange(x, ...)

yrange(x, ...)

\method{xrange}{default}(x)

\method{yrange}{default}(x)

local_triangulate(x, family = "laea")
}
\arguments{
\item{x}{gris object}
}
\value{
numeric
}
\description{
min/max for gris
}

