% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gris.R
\name{gris}
\alias{[.gris}
\alias{as.gris}
\alias{gris}
\alias{plot.gris}
\alias{print.gris}
\title{gris}
\usage{
gris(x, ...)

\method{[}{gris}(x, i, j, drop = FALSE)

\method{print}{gris}(x, ..., n = NULL, width = NULL)

\method{plot}{gris}(x, y, triangles = FALSE, ...)

as.gris(x, ...)
}
\arguments{
\item{x}{Spatial* object}

\item{...}{unused}

\item{i}{indices for extract}

\item{j}{indices for extract}

\item{drop}{ignored}

\item{n}{n}

\item{width}{width}

\item{y}{y}
}
\value{
gris
}
\description{
gris
}
\section{Methods}{


\code{gris} implements these base methods:

\describe{
\item{plot}{Plots the object according to its topology type, or by a chosen \code{type}. This
method follows base plotting with types "p" points, "l" lines, and adds "pp" for polygons}
\item{print}{Prints the object and vertices tables with a summary of the topology linkages}
\item{\code{[}}{Subset as an object table on attributes, propagates down through the geometry
links to return a subsetted gris object}
}
}

