% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gris.R
\name{grisFromFortify}
\alias{grisFromFortify}
\title{Not yet implemented}
\usage{
grisFromFortify(x)
}
\arguments{
\item{x}{}
}
\value{
gris
}
\description{
Not yet implemented
}
\examples{
\dontrun{
 data(wrld_simpl)
 g <- gris(wrld_simpl)
 sp <- as.SpatialPolygonsDataFrame(g)
 gm <- raster::geom(sp)
}
}

