% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign.R
\name{grisToSPbase}
\alias{grisToSPbase}
\title{Return a list of Polygon, Line or matrix from gris}
\usage{
grisToSPbase(x, type = "pp")
}
\arguments{
\item{x}{gris object}

\item{type}{return type:  pp-oly, l-ine, p-oints}
}
\description{
Return a list of Polygon, Line or matrix from gris
}

