% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gris.R
\name{mkpslg}
\alias{mkpslg}
\title{Make a Planar Straight Line Graph from gris.}
\usage{
mkpslg(x)
}
\arguments{
\item{x}{gris object}
}
\value{
pslg object from RTriangle
}
\description{
Make a Planar Straight Line Graph from gris.
}

