% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gris.R
\name{pl}
\alias{pl}
\title{Simple plot for vertices by branch and object}
\usage{
pl(x, col = NULL, debug = FALSE, asp = NULL, ..., type = "p")
}
\arguments{
\item{x}{vertex table}

\item{col}{optional colours for each object}

\item{asp}{aspect ratio for plot}

\item{...}{arguments based to graphics::plot}

\item{type}{p or l for poly or line}
}
\value{
\code{NULL}
}
\description{
Input is a data_frame of x, y, coordinates and .br0 and .ob0 the indices of individual branches and objects.
}

