% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trianglePrimitives.R
\name{pointInTriangle}
\alias{pointBary}
\alias{pointInTriangle}
\title{Point in triangle test.}
\usage{
pointInTriangle(x, pt)

pointBary(x, pt)
}
\arguments{
\item{x}{gris object with triangulation}

\item{pt}{matrix of xy coordinates}
}
\value{
for PointInTriangle, the index of triangle for each point, in .vx0 form
for PointBary, the table of barycentric coordinates and .vx0 index
}
\description{
Point in triangle test.
}

