% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rgl.R
\name{pquads}
\alias{pquads}
\title{pquads}
\usage{
pquads(x, texture = NULL, texcoords = NULL, subset = NULL, ...)
}
\arguments{
\item{x}{mesh object}

\item{texture}{path to PNG file}

\item{texcoords}{texture coordinates}
}
\description{
pquads
pquads
}

