% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj4.R
\name{ptransform}
\alias{ptransform}
\alias{ptransform.gris}
\title{Map projection for gris}
\usage{
ptransform(x, ...)

\method{ptransform}{gris}(x, target, ...)
}
\arguments{
\item{x}{gris object}

\item{...}{passed to methods}

\item{target}{proj4 string}
}
\value{
gris object
}
\description{
Map projection for gris
}
\examples{
library(maptools)
data(wrld_simpl)
g <- gris(wrld_simpl[10, ])
ptransform(g, "+proj=laea +ellps=WGS84")
}

