% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.R
\name{quadmeshFromRaster}
\alias{quadmeshFromRaster}
\title{Create a quad-type mesh for use in rgl}
\usage{
quadmeshFromRaster(x, z = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{raster object for mesh structure}

\item{z}{raster object for height values}

\item{na.rm}{remove quads where missing values?}
}
\value{
mesh3d
}
\description{
Convert a \code{\link[raster]{RasterLayer}} to a \code{\link[rgl]{mesh3d}} quad-mesh.
}

