% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.R
\name{ras2gris}
\alias{ras2gris}
\title{Raster to gris object}
\usage{
ras2gris(x, z = NULL)
}
\arguments{
\item{x}{RasterLayer}

\item{z}{RasterLayer to extract Z value from}
}
\value{
gris
}
\description{
Raster to gris object
}

