% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gris.R
\name{sbs}
\alias{sbs}
\title{Subset for a vertex/branch/object object}
\usage{
sbs(x, subset, ...)
}
\arguments{
\item{x}{list of \code{v} vertex, \code{b} branch and \code{o} object tables}

\item{subset}{a subset of the \code{o} object}

\item{...}{ignored}
}
\value{
list of \code{v} vertex, \code{b} branch and \code{o} object tables
}
\description{
Subset for a vertex/branch/object object
}

