% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gris.R
\name{gris.SpatialPolygonsDataFrame}
\alias{gris.SpatialLinesDataFrame}
\alias{gris.SpatialPointsDataFrame}
\alias{gris.SpatialPolygonsDataFrame}
\title{Convert Spatial* objects to gris}
\usage{
\method{gris}{SpatialPolygonsDataFrame}(x, ...)

\method{gris}{SpatialLinesDataFrame}(x, ...)

\method{gris}{SpatialPointsDataFrame}(x, ...)
}
\arguments{
\item{x}{Spatial* object}

\item{...}{not used}
}
\value{
gris
}
\description{
Description of it
}
\details{
Details are
}

