% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trianglePrimitives.R
\name{triangulate}
\alias{triangulate}
\alias{triangulate.gris}
\title{Add triangle primitives}
\usage{
triangulate(x, ...)
}
\arguments{
\item{x}{gris object}

\item{...}{passed on to \code{\link[Rtriangle]{triangulate}}, as arguments after the input \code{\link[Rtriangle]{pslg}}}
}
\description{
Add triangle primitives
}
\details{
only \code{a} has been extensively explored, this is the minimum triangle area in the units of the coordinates
}
\examples{
\donttest{
\dontrun{
library(gris)
library(maptools)
data(wrld_simpl)
gall <- gris(wrld_simpl)
g <- gall[gall$o$NAME \%in\% c("Australia", "New Zealand"), ]
g$v <- g$v \%>\% mutate(x = ifelse(x < 0, x + 360, x))

gT <- gris::triangulate(g)
plot(g)
plot(gT, triangles = TRUE, add = TRUE)
}}

}

