% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trianglePrimitives.R
\name{triGris}
\alias{triGris}
\title{Add triangle primitives}
\usage{
triGris(x)
}
\arguments{
\item{x}{gris object}
}
\description{
Add triangle primitives
}
\examples{
\donttest{
\dontrun{
library(gris)
library(maptools)
data(wrld_simpl)
gall <- gris(wrld_simpl)
g <- gall[gall$o$NAME \%in\% c("Australia", "New Zealand"), ]
g$v <- g$v \%>\% mutate(x = ifelse(x < 0, x + 360, x))
library(RTriangle)
gT <- triGris(g)
plot(g)
plotT(gT)
}}
}

