% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.R
\name{bgl}
\alias{bgl}
\title{Title}
\usage{
bgl(x, z = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{raster object for mesh structure}

\item{z}{raster object for height values}

\item{na.rm}{remove quads where missing values?}
}
\value{
mesh3d
}
\description{
Title
}

