% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign.R
\name{as.SpatialPolygonsDataFrame}
\alias{as.SpatialPolygonsDataFrame}
\alias{as.SpatialPolygonsDataFrame.gris}
\title{Return a list of Polygon, Line or matrix from gris}
\usage{
as.SpatialPolygonsDataFrame(x, ...)

\method{as.SpatialPolygonsDataFrame}{gris}(x, ...)
}
\arguments{
\item{x}{gris object}

\item{...}{ignored}

\item{type}{return type:  pp-oly, l-ine, p-oints}

\item{x}{gris object}
}
\description{
Return a list of Polygon, Line or matrix from gris

Spatial classes are defined in the \link{sp} package.
}

