% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign.R
\name{grisToSPbase}
\alias{grisToSPbase}
\title{Return a list of Polygon, Line or matrix}
\usage{
grisToSPbase(x, type = "pp")
}
\arguments{
\item{x}{gris}

\item{type}{pp-oly, l-ine, p-oints}
}
\description{
Return a list of Polygon, Line or matrix
}

