% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.R
\name{pquads}
\alias{pquads}
\title{pquads}
\usage{
pquads(x, texture = NULL, texcoords = NULL, subset = NULL, ...)
}
\arguments{
\item{x}{mesh object}

\item{texture}{path to PNG file}

\item{texcoords}{texture coordinates}

\item{subset}{option index (on ib)}

\item{...}{arguments pass to \code{\link[rgl]{rgl.quads}}}
}
\description{
pquads
}

