% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gris.R
\name{gris.SpatialPolygonsDataFrame}
\alias{gris.Spatial}
\alias{gris.SpatialLinesDataFrame}
\alias{gris.SpatialMultiPointsDataFrame}
\alias{gris.SpatialPointsDataFrame}
\alias{gris.SpatialPolygonsDataFrame}
\title{Convert Spatial*DataFrame objects to gris}
\usage{
\method{gris}{SpatialPolygonsDataFrame}(x, ...)

\method{gris}{SpatialLinesDataFrame}(x, ...)

\method{gris}{SpatialPointsDataFrame}(x, ...)

\method{gris}{SpatialMultiPointsDataFrame}(x, ...)
}
\arguments{
\item{x}{Spatial* object}

\item{...}{not used}
}
\value{
gris
}
\description{
All availables Spatial*DataFrame types are supported, and are interpreted via the Branch model. 
The gris function can ingest \code{\link[sp]{SpatialPolygonsDataFrame}}, \code{\link[sp]{SpatialLinesDataFrame}}, \code{\link[sp]{SpatialPointsDataFrame}}, and \code{\link[sp]{SpatialMultiPointsDataFrame}} objects.
}
\details{
See  \code{vignette("branch-vs-primitives")}
}

